+++ lib/pcb/svg-object-fns/create-svg-objects-from-pcb-note-text.ts
@@ -4,7 +4,7 @@ import { debugPcb } from "lib/utils/debug"
 import { applyToPoint } from "transformation-matrix"
 import type { PcbContext } from "../convert-circuit-json-to-pcb-svg"
 import { colorMap } from "lib/utils/colors"
-import { getSvgTextAnchorAlignment } from "../text-anchor-alignment"
+import { getSvgTextAnchorAlignment } from "../get-svg-text-anchor-alignment"
 
 const DEFAULT_OVERLAY_COLOR = colorMap.board.user_2
